+++ tests/dummy/app/routes/components-acceptance-tests/flexberry-objectlistview/custom-filter.js
@@ -2,7 +2,7 @@ import ListFormRoute from 'ember-flexberry/routes/list-form';
 
 import { Query } from 'ember-flexberry-data';
 
-const { SimplePredicate, StringPredicate } = Query;
+const { SimplePredicate, StringPredicate, DatePredicate } = Query;
 
 export default ListFormRoute.extend({
   /**
@@ -56,6 +56,10 @@ export default ListFormRoute.extend({
 
     if (filter.type === 'decimal') {
       return new SimplePredicate(filter.name, filter.condition, filter.pattern ? Number(filter.pattern) : filter.pattern);
+    }
+
+    if (filter.type === 'date') {
+      return new DatePredicate(filter.name, filter.condition, filter.pattern, true);
     }
 
     return this._super(...arguments);